test/requirements.txt
paramiko
mysql-connector-python
sqlmap
requests
pymodbus
bacpypes3
cpppo
python-snap7
pysnmp
